properly implement host and runtime regex
